Use u32 instead of u64

+++ gpu/src/metal/shaders/fields/bls12381.h.metal
@@ -63,7 +63,7 @@ template [[ host_name("calculate_points_sum") ]]
 [[kernel]] void calculate_points_sum<FpBLS12381, BLS12381>(
     constant FpBLS12381*,
     constant BLS12381*,
-    constant uint64_t&,
+    constant uint32_t&,
     constant uint32_t&,
     device BLS12381*
 );
@@ -71,7 +71,7 @@ template [[ host_name("calculate_points_sum") ]]
 template [[ host_name("calculate_window") ]]
 [[kernel]] void calculate_window<BLS12381>(
     device BLS12381*,
-    constant uint64_t&,
+    constant uint32_t&,
     device BLS12381*,
     device BLS12381&
 );
@@ -79,6 +79,6 @@ template [[ host_name("calculate_window") ]]
 template [[ host_name("reduce_windows") ]]
 [[kernel]] void reduce_windows<BLS12381>(
     constant BLS12381*,
-    constant uint64_t&,
+    constant uint32_t&,
     device BLS12381&
 );
